move complete profile alert
